% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gii.R
\name{compute_gii}
\alias{compute_gii}
\title{Compute Gender Inequality Index (GII) and Select Bottom 15 Countries}
\usage{
compute_gii(data)
}
\arguments{
\item{data}{A data frame containing the required metrics for GII computation.}
}
\value{
A data frame of the bottom 15 countries sorted by GII values.
}
\description{
This is an internal helper function to compute the GII for each country in the data
and then select the bottom 15 countries based on their GII values.
}
\keyword{internal}
