% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_inequality_index.R
\name{gender_inequality_index}
\alias{gender_inequality_index}
\title{Gender Inequality Index (GII)}
\usage{
gender_inequality_index(data)
}
\arguments{
\item{data}{A data frame containing the required metrics for GII computation.}
}
\value{
A numeric vector representing the GII values.
}
\description{
Computes the Gender Inequality Index (GII) based on the provided dataset.
}
\examples{

data(real_data_GII) # Load example dataset
gii_results <- gender_inequality_index(real_data_GII)
print(gii_results)

}
