% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupational_segregation.R
\name{occupational_segregation}
\alias{occupational_segregation}
\title{Occupational Segregation Indices}
\usage{
occupational_segregation(data)
}
\arguments{
\item{data}{A data frame containing the required metrics for occupational segregation computation.}
}
\value{
A numeric vector representing the Occupational Segregation Indices.
}
\description{
Computes the Occupational Segregation Indices based on the provided dataset.
}
\examples{
\dontrun{
data(occupational_segregation_data) # Load example dataset
os_results <- occupational_segregation(occupational_segregation_data)
print(os_results)
}

}
