% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gdi.R
\name{plot_gdi}
\alias{plot_gdi}
\title{Plot Gender Development Index (GDI) for Bottom 15 Countries}
\usage{
plot_gdi(data)
}
\arguments{
\item{data}{A data frame containing the required metrics for GDI computation.}
}
\value{
A ggplot2 object visualizing the GDI for the bottom 15 countries.
}
\description{
Plot Gender Development Index (GDI) for Bottom 15 Countries
}
\examples{
data(real_data_GDI) # Load example dataset
plot_gdi(real_data_GDI)
}
