% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gem.R
\name{plot_gem}
\alias{plot_gem}
\title{Plot Gender Empowerment Measure (GEM) for Top 15 Countries}
\usage{
plot_gem(data)
}
\arguments{
\item{data}{A data frame containing the required metrics for GEM visualization.}
}
\value{
A ggplot object representing the bar chart of the GEM and other metrics.
}
\description{
Visualizes the Gender Empowerment Measure (GEM) along with other metrics
such as female parliament seats, female professional positions, and
female to male earned income ratio for the top 15 countries based on GEM.
}
\examples{

data(real_data_GEM) # Load example dataset
plot_gem(real_data_GEM)

}
