% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gii.R
\name{plot_gii}
\alias{plot_gii}
\title{Plot Gender Inequality Index (GII) for Bottom 15 Countries}
\usage{
plot_gii(data)
}
\arguments{
\item{data}{A data frame containing the required metrics for GII computation.}
}
\value{
A ggplot2 object visualizing the GII for the bottom 15 countries.
}
\description{
Plot Gender Inequality Index (GII) for Bottom 15 Countries
}
\examples{
data(real_data_GII) # Load example dataset
plot_gii(real_data_GII)
}
