% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gii.R
\name{plot_gii_education}
\alias{plot_gii_education}
\title{Plot Secondary Education Participation for Bottom 15 Countries by Gender}
\usage{
plot_gii_education(data)
}
\arguments{
\item{data}{A data frame containing the required metrics for GII computation.}
}
\value{
A ggplot2 object visualizing the secondary education participation for males and females in the bottom 15 countries.
}
\description{
Plot Secondary Education Participation for Bottom 15 Countries by Gender
}
\examples{
data(real_data_GII) # Load example dataset
plot_gii_education(real_data_GII)
}
