% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gpg.R
\name{plot_gpg}
\alias{plot_gpg}
\title{Plot Gender Pay Gap for Top 5 Countries}
\usage{
plot_gpg(data)
}
\arguments{
\item{data}{A data frame containing the required metrics for GPG computation.}
}
\value{
A ggplot2 object visualizing the GPG for the top 5 countries.
}
\description{
Visualizes the Gender Pay Gap (GPG) for the top 5 countries based on their GPG.
}
\examples{
data(real_data_GPG) # Load example dataset
plot_gpg(real_data_GPG)


}
