% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gpg.R
\name{plot_gpg_q2}
\alias{plot_gpg_q2}
\title{Plot Gender Pay Gap for Second Quartile Countries}
\usage{
plot_gpg_q2(data)
}
\arguments{
\item{data}{A data frame containing the required metrics for GPG computation.}
}
\value{
A ggplot2 object visualizing the GPG for Q2 countries.
}
\description{
Visualizes the Gender Pay Gap (GPG) for countries in the second quartile (Q2) based on their GPG.
}
\examples{
data(real_data_GPG) # Load example dataset
plot_gpg_q2(real_data_GPG)

}
