% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_inequality_index.R
\docType{data}
\name{real_data_GII}
\alias{real_data_GII}
\title{Original Data for Gender Inequality Index (GII)}
\format{
A data frame with 191 rows and 8 columns:
\describe{
  \item{country}{A character vector representing the name of the country.}
  \item{maternal_mortality_ratio}{A numeric vector representing the maternal mortality ratio.}
  \item{adolescent_birth_rate}{A numeric vector representing the adolescent birth rate.}
  \item{female_parliament_seats}{A numeric vector representing the proportion of parliamentary seats occupied by females in the respective country. Measured as a proportion (0-1).}
  \item{female_secondary_education}{A numeric vector representing the proportion of females with secondary education in the respective country. Measured as a proportion (0-1).}
  \item{male_secondary_education}{A numeric vector representing the proportion of males with secondary education in the respective country. Measured as a proportion (0-1).}
  \item{female_labor_force}{A numeric vector representing the proportion of females in the labor force in the respective country. Measured as a proportion (0-1).}
  \item{male_labor_force}{A numeric vector representing the proportion of males in the labor force in the respective country. Measured as a proportion (0-1).}
}
}
\source{
Data obtained from \href{https://www.kaggle.com/dsv/6359326}{Kaggle}.
}
\usage{
data(real_data_GII)
}
\description{
A dataset containing observed values for the GII analysis. This data includes metrics for evaluating gender disparities in reproductive health, empowerment, and labor market participation across different countries.
}
