% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_pay_gap.R
\docType{data}
\name{real_data_GPG}
\alias{real_data_GPG}
\title{Original Data for Gender Pay Gap (GPG)}
\format{
A data frame with 191 rows and 3 columns:
\describe{
  \item{country}{A character vector representing the name of the country.}
  \item{female_income}{A numeric vector representing the per capita female National Income in the respective country.}
  \item{male_income}{A numeric vector representing the per capita male National Income in the respective country.}
}
}
\source{
Data obtained from \href{https://www.kaggle.com/dsv/6359326}{Kaggle}.
}
\usage{
data(real_data_GPG)
}
\description{
A dataset containing observed values for the GPG analysis. This data includes metrics for evaluating income disparities between genders across different countries.
}
