% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_inequality_index.R
\docType{data}
\name{simulated_data_GII}
\alias{simulated_data_GII}
\title{Simulated Data for Gender Inequality Index (GII)}
\format{
A data frame with 200 rows and 6 columns:
\describe{
  \item{country}{A character vector representing the name of the country.}
  \item{female_reproductive_age}{A numeric vector representing the reproductive age for females in the respective country.}
  \item{male_reproductive_age}{A numeric vector representing the reproductive age for males in the respective country.}
  \item{female_secondary_education}{A numeric vector representing the secondary education completion rate for females in the respective country. Measured as a proportion (0-1).}
  \item{male_secondary_education}{A numeric vector representing the secondary education completion rate for males in the respective country. Measured as a proportion (0-1).}
  \item{female_labor_force}{A numeric vector representing the labor force participation rate for females in the respective country. Measured as a proportion (0-1).}
}
}
\source{
The dataset is simulated for the purpose of this package.
}
\usage{
data(simulated_data_GII)
}
\description{
A dataset containing simulated values for the GII analysis.
}
