% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_pay_gap.R
\docType{data}
\name{simulated_data_GPG}
\alias{simulated_data_GPG}
\title{Simulated Data for Gender Pay Gap (GPG)}
\format{
A data frame with 200 rows and 3 columns:
\describe{
  \item{country}{A character vector representing the name of the country.}
  \item{female_salary}{A numeric vector representing the average salary of females in the respective country. Measured in international dollars.}
  \item{male_salary}{A numeric vector representing the average salary of males in the respective country. Measured in international dollars.}
}
}
\source{
The dataset is simulated for the purpose of this package.
}
\usage{
data(simulated_data_GPG)
}
\description{
A dataset containing simulated values for the GPG analysis.
}
