% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupational_segregation.R
\docType{data}
\name{simulated_data_OS}
\alias{simulated_data_OS}
\title{Simulated Data for Occupational Segregation}
\format{
A data frame with 200 rows and 3 columns:
\describe{
  \item{country}{A character vector representing the name of the country.}
  \item{female_occupation}{A character vector representing the most common occupation for females in the respective country.}
  \item{male_occupation}{A character vector representing the most common occupation for males in the respective country.}
}
}
\source{
The dataset is simulated for the purpose of this package.
}
\usage{
data(simulated_data_OS)
}
\description{
A dataset containing simulated values for Occupational Segregation analysis.
}
