% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_use_analysis.R
\docType{data}
\name{simulated_data_TUA}
\alias{simulated_data_TUA}
\title{Simulated Data for Time-use Analysis}
\format{
A data frame with 200 rows and 4 columns:
\describe{
  \item{person}{ID of person.}
  \item{gender}{Identified gender of the person.}
  \item{unpaid_work_hours}{Unpaid work hours}
  \item{paid_work_hours}{Paid work hours}
}
}
\source{
The dataset is simulated for the purpose of this package.
}
\usage{
data(simulated_data_TUA)
}
\description{
A dataset containing simulated values for Time-use Analysis.
}
