% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gii.R
\name{plot_gii_empowerment}
\alias{plot_gii_empowerment}
\title{Plot Female Parliament Seats for Top 15 Countries}
\usage{
plot_gii_empowerment(data)
}
\arguments{
\item{data}{A data frame containing the required metrics for GII computation.}
}
\value{
A ggplot2 object visualizing female parliament seats for the top 15 countries.
}
\description{
Plot Female Parliament Seats for Top 15 Countries
}
\examples{
data(real_data_GII) # Load example dataset
plot_gii_empowerment(real_data_GII)
}
