% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_pay_gap.R
\docType{data}
\name{real_data_GPG}
\alias{real_data_GPG}
\title{Original Data for Gender Pay Gap (GPG)}
\format{
A data frame with 191 rows and 3 columns:
\describe{
  \item{country}{Name of Country}
  \item{female_income}{Per capita female National Income}
  \item{male_income}{Per capita male National Income}
}
}
\source{
Data obtained from the following sources:
\itemize{
  \item{\href{https://hdr.undp.org/data-center/documentation-and-downloads}{UNDP Human Development Reports Data Center}}
  \item{\href{https://genderdata.worldbank.org/indicators/}{World Bank Gender Data Portal}}
}
}
\usage{
data(real_data_GPG)
}
\description{
A dataset containing observed values for the GPG analysis.
}
