% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupational_segregation.R
\docType{data}
\name{simulated_data_OS}
\alias{simulated_data_OS}
\title{Simulated Data for Occupational Segregation}
\format{
A data frame with 200 rows and 3 columns:
\describe{
  \item{occupation}{Occupation of the person.}
  \item{male_count}{Male Count.}
  \item{female_count}{Female Count.}
}
}
\source{
The dataset is simulated for the purpose of this package.
}
\usage{
data(simulated_data_OS)
}
\description{
A dataset containing simulated values for Occupational Segregation analysis.
}
